refactor: 패키지 구조를 core, web, console로 분리
