+++ src/main/java/roomescape/core/domain/Name.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.core.domain;
 
 import java.util.Objects;
