+++ src/main/java/roomescape/core/repository/H2ReservationRepository.java
@@ -1,11 +1,11 @@
-package roomescape.repository;
+package roomescape.core.repository;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Name;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
+import roomescape.core.domain.Name;
+import roomescape.core.domain.Reservation;
+import roomescape.core.domain.ReservationTime;
 
 import javax.sql.DataSource;
 import java.util.List;
