+++ src/main/java/roomescape/web/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
