+++ src/main/java/roomescape/web/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.web;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -8,12 +8,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.controller.request.ReservationRequest;
-import roomescape.controller.response.ReservationResponse;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ReservationTimeRepository;
+import roomescape.core.domain.Reservation;
+import roomescape.core.domain.ReservationTime;
+import roomescape.web.request.ReservationRequest;
+import roomescape.web.response.ReservationResponse;
+import roomescape.core.repository.ReservationRepository;
+import roomescape.core.repository.ReservationTimeRepository;
 
 import java.net.URI;
 import java.util.List;
