+++ src/main/java/roomescape/web/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.web;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -8,10 +8,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.ReservationTime;
-import roomescape.controller.request.ReservationTimeRequest;
-import roomescape.controller.response.ReservationTimeResponse;
-import roomescape.repository.ReservationTimeRepository;
+import roomescape.core.domain.ReservationTime;
+import roomescape.web.request.ReservationTimeRequest;
+import roomescape.web.response.ReservationTimeResponse;
+import roomescape.core.repository.ReservationTimeRepository;
 
 import java.util.List;
