+++ src/main/java/roomescape/web/response/ReservationTimeResponse.java
@@ -1,7 +1,7 @@
-package roomescape.controller.response;
+package roomescape.web.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import roomescape.domain.ReservationTime;
+import roomescape.core.domain.ReservationTime;
 
 import java.time.LocalTime;
