+++ src/test/java/roomescape/core/repository/H2ReservationRepositoryTest.java
@@ -1,13 +1,15 @@
-package roomescape.repository;
+package roomescape.core.repository;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.Name;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
+import roomescape.core.domain.Name;
+import roomescape.core.domain.Reservation;
+import roomescape.core.domain.ReservationTime;
+import roomescape.core.repository.H2ReservationRepository;
+import roomescape.core.repository.H2ReservationTimeRepository;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
