+++ src/test/java/roomescape/web/acceptance/BasicAcceptanceTest.java
@@ -1,4 +1,4 @@
-package roomescape.acceptance;
+package roomescape.web.acceptance;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
